src/tests/test_aesthetics.py
import unittest
from src import aesthetics

class TestAesthetics(unittest.TestCase):

    def setUp(self):
        self.dataSet = "mock_data"
        self.userInput = "mock_input"
        self.encryptedData = "mock_encrypted_data"
        self.authenticatedUser = "mock_authenticated_user"

    def test_applyAesthetics(self):
        result = aesthetics.applyAesthetics(self.dataSet, self.userInput, self.encryptedData, self.authenticatedUser)
        self.assertIsNotNone(result)

    def test_optimizeUsability(self):
        result = aesthetics.optimizeUsability(self.dataSet, self.userInput, self.encryptedData, self.authenticatedUser)
        self.assertIsNotNone(result)

    def test_applyJournalismPrinciples(self):
        result = aesthetics.applyJournalismPrinciples(self.dataSet, self.userInput, self.encryptedData, self.authenticatedUser)
        self.assertIsNotNone(result)

    def test_applyMusicTheory(self):
        result = aesthetics.applyMusicTheory(self.dataSet, self.userInput, self.encryptedData, self.authenticatedUser)
        self.assertIsNotNone(result)

if __name__ == '__main__':
    unittest.main()